+++ bin/check-es-heap.rb
@@ -107,13 +107,12 @@ class ESHeap < Sensu::Plugin::Check::CLI
   end
 
   def acquire_heap_data(return_max = false)
-    if Gem::Version.new(acquire_es_version) >= Gem::Version.new('1.0.0')
-      stats = acquire_es_resource('/_nodes/_local/stats?jvm=true')
-      node = stats['nodes'].keys.first
-    else
-      stats = acquire_es_resource('/_cluster/nodes/_local/stats?jvm=true')
-      node = stats['nodes'].keys.first
-    end
+    node = stats['nodes'].keys.first
+    stats = if Gem::Version.new(acquire_es_version) >= Gem::Version.new('1.0.0')
+              acquire_es_resource('/_nodes/_local/stats?jvm=true')
+            else
+              acquire_es_resource('/_cluster/nodes/_local/stats?jvm=true')
+            end
     begin
       if return_max
         return stats['nodes'][node]['jvm']['mem']['heap_used_in_bytes'], stats['nodes'][node]['jvm']['mem']['heap_max_in_bytes']